feat(ui): extract withdraw logic into CamWithdraw component and add address pixel icons (#332)
* refactor: create new component for withdraw

* refactor(MyMessenger): remove balances from my messenger view

* feat(MyMessenger): add icon pixel in my messeger view
