+++ package.json
@@ -89,6 +89,7 @@
     "dependencies": {
         "clsx": "^2.1.1",
         "react": "^18.2.0",
+        "react-blockies": "^1.4.1",
         "react-circle-flags": "^0.0.18",
         "react-dom": "^18.2.0"
     }
